+++ src/main/java/org/panteleyev/money/TransactionTableView.java
@@ -56,7 +56,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
@@ -181,7 +180,6 @@ public class TransactionTableView extends TableView<Transaction> {
         TableColumn<Transaction, CheckBox> approvedColumn = new TableColumn<>("");
 
         approvedColumn.setCellValueFactory(p -> {
-
             CheckBox cb = new CheckBox();
 
             Transaction value = p.getValue();
@@ -199,7 +197,7 @@ public class TransactionTableView extends TableView<Transaction> {
         });
         approvedColumn.setSortable(false);
 
-        getColumns().setAll(Arrays.asList(dayColumn,
+        getColumns().setAll(List.of(dayColumn,
                 typeColumn,
                 accountFromColumn,
                 accountCreditedColumn,