+++ src/main/java/org/panteleyev/money/statements/StatementParser.java
@@ -29,6 +29,15 @@ package org.panteleyev.money.statements;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVRecord;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.NodeList;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathConstants;
+import javax.xml.xpath.XPathExpression;
+import javax.xml.xpath.XPathFactory;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
@@ -41,6 +50,7 @@ import java.util.List;
 
 public final class StatementParser {
     private static final DateTimeFormatter RBA_DATE_FORMAT = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm");
+    private static final DateTimeFormatter OFX_DATE_FORMAT = DateTimeFormatter.ofPattern("yyyyMMddHHmmss");
     private static final char RBA_DELIMITER = ';';
     private static final int RBA_RECORD_SIZE = 10;
     private static final String RBA_ENCODING = "windows-1251";
@@ -49,6 +59,9 @@ public final class StatementParser {
         switch (type) {
             case RAIFFEISEN_CREDIT_CARD_CSV:
                 return parseRaiffeisenCSV(inStream);
+
+            case RAIFFEISEN_CARD_OFX:
+                return parseRaiffeisenOfx(inStream);
 
             default:
                 throw new IllegalArgumentException();
@@ -85,6 +98,57 @@ public final class StatementParser {
         } catch (IOException ex) {
             throw new UncheckedIOException(ex);
         } catch (DateTimeParseException ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+
+    private static Statement parseRaiffeisenOfx(InputStream inStream) {
+        try {
+            List<StatementRecord> records = new ArrayList<>();
+            DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+            DocumentBuilder documentBuilder = factory.newDocumentBuilder();
+            Document doc = documentBuilder.parse(inStream);
+
+            XPathFactory xPathFactory = XPathFactory.newInstance();
+            XPath xPath = xPathFactory.newXPath();
+            XPathExpression expression = xPath.compile("/OFX/BANKMSGSRSV1/STMTTRNRS/STMTRS");
+
+            NodeList nl = (NodeList) expression.evaluate(doc, XPathConstants.NODESET);
+
+            for (int i = 0; i < nl.getLength(); i++) {
+                Element stmtrs = (Element)nl.item(i);
+                NodeList transactions = stmtrs.getElementsByTagName("STMTTRN");
+
+                StatementRecord.Builder builder = new StatementRecord.Builder();
+                for (int j = 0; j < transactions.getLength(); j++) {
+                    Element transaction = (Element)transactions.item(j);
+                    NodeList children = transaction.getChildNodes();
+
+                    for (int k = 0; k < children.getLength(); k++) {
+                        Element field = (Element)children.item(k);
+
+                        switch (field.getTagName()) {
+                            case "NAME":
+                                builder.counterParty(field.getTextContent());
+                                break;
+                            case "DTPOSTED":
+                                builder.actual(LocalDate.parse(field.getTextContent(), OFX_DATE_FORMAT));
+                                break;
+                            case "DTAVAIL":
+                                builder.execution(LocalDate.parse(field.getTextContent(), OFX_DATE_FORMAT));
+                                break;
+                            case "TRNAMT":
+                                builder.amount(field.getTextContent());
+                                break;
+                        }
+                    }
+
+                    records.add(builder.build());
+                }
+            }
+
+            return new Statement(Statement.StatementType.RAIFFEISEN_CARD_OFX, records);
+        } catch (Exception ex) {
             throw new RuntimeException(ex);
         }
     }